refactor to move script outside loginform
